delete the old components #227
I deleted the old progress stepper and the last new one and fixed the style so it works with the new style guide
